raising exceptions in python-worker for http error code from r-worker (#98)
The tasks in the python worker rely on using r.json() to read the responses of the requests made to the r-worker. The method r.json() will throw an exception for error codes because they don't contain a body. 

The error `json.decoder.JSONDecodeError: Extra data: line 1 column 5 (char 4)` is not very descriptive, so for the time being I added the `raise_for_status()` so we will have clear exceptions for status codes > 400. In the future status codes should probably be handled more clearly.