Remove default experiment ID
As far as I can tell this is a vestige of when
this experiment was the common test data. Now we
don't test with only this (and it is not included
in the repo), it is a misleading default. With this
change, `docker-compose up` without setting
EXPERIMENT_ID will print `WARNING: The EXPERIMENT_ID
variable is not set. Defaulting to a blank string.`,
like it does for other important variables.